requirements.txt
Django>=5.1.5,<6.0
pillow
django-autoslug
django-unfold
gunicorn 
whitenoise
django-environ
dj-database-url
psycopg2-binary
django-cloudinary-storage
django-compressor
django-quill-editor